requirements.txt
accelerate==0.24.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
diffusers==0.22.1
einops==0.7.0
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
huggingface-hub==0.19.4
idna==3.6
imageio==2.31.6
imageio-ffmpeg==0.4.9
imagesize==1.4.1
imbalanced-learn==0.10.1
imgaug==0.4.0
importlib-metadata==7.0.0
imutils==0.5.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.2
pandas==1.5.3
Pillow==10.0.1
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.15.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
typing_extensions==4.9.0
urllib3==2.0.7
zipp==3.17.0
